+++ packages/blocks/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/blocks
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.0
+
+## 0.1.0
+
+### Minor Changes
+
+- Fixes
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.6.0
+
 ## 0.0.13
 
 ### Patch Changes